Convert iIMDShape.min/max from Vector3f to Vector3i, since they are used in the game state.
Don't know if that could cause any desynchs, but better not to risk it.

+++ lib/ivis_common/ivisdef.h
@@ -77,7 +77,7 @@ typedef struct _iIMDShape {
 	int texpage;
 	int tcmaskpage;
 	int sradius, radius;
-	Vector3f min, max;
+	Vector3i min, max;
 
 	Vector3f ocen;
 	unsigned short numFrames;
